Add examples of driving GPIO pins.
The Blinkt! hardware consists of 8 RGB LEDs which can be controlled
through the GPIO pins. Data needs to be clocked into the LED's
controller and the result is that the lights flash up. It's not
fast on RISC OS Pyromaniac over slow USB controlled GPIO pins, but
it probably works quite nicely on real hardware.