Updated the comments for the PCF8591 board for the jumpers.
The PCF8591 board that I have includes an LDR, Thermistor and a
potentiometer. These are wired to some of the analogue inputs when
the jumpers are in place. If you try to use the analogue inputs
without removing the jumpers, you'll get variable results because
they're interfering with one another!

There is now a comment describing this.